split __init__ and simCLRDatasetWrapper
